+++ examples/invoice-helper/miniprogram/components/amount-chip/amount-chip.js
@@ -1,3 +1,0 @@
-Component(
-  require("../../engine/moon-engine.js").component("components/amount-chip/amount-chip")
-);